delete file commnets.js
